requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2
newsapi-python==0.2.7
python-dateutil==2.8.2
plotly==5.19.0
watchdog==4.0.0 